Added customization and notification on reset.
Notification currently only works on Linux when using Dunst. It is
possible by setting an icon for the notification using the optional
--raw_icon argument when using the reset/stop command.

Prefix customization is possible using the --prefix argument when using
the status command.

+++ src/main/java/io/hoogland/Main.java
@@ -2,22 +2,16 @@ package io.hoogland;
 
 import io.hoogland.models.Stopwatch;
 import io.hoogland.utils.FileUtils;
+import io.hoogland.utils.OSUtils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
 public class Main {
-    /**
-     * Prefix icon printed in front of the timer/state.
-     */
-    public static final String PREFIX = "\uE151";
 
     /**
      * Temp directory used to store stopwatch objects.
@@ -45,6 +39,19 @@ public class Main {
         // Cast arguments to lowercase for easier matching.
         List<String> arguments = Arrays.stream(args).map(String::toLowerCase).toList();
 
+        // Set fallback icon for prefix.
+        String prefix = "\uE151";
+
+        // Check if prefix argument is given, if so change it to given prefix.
+        if (arguments.stream().anyMatch(arg -> arg.toLowerCase().contains("--prefix="))) {
+            // Find the String in the list that has the prefix.
+            for (String arg : arguments) {
+                if (arg.toLowerCase().startsWith("--prefix=")) {
+                    prefix = arg.substring(arg.indexOf("=") + 1);
+                }
+            }
+        }
+
         try {
             if (arguments.contains("start")) {
                 // If start is an argument, start a new stopwatch and write it to file.
@@ -56,33 +63,49 @@ public class Main {
                 if (stopwatchFile.exists()) {
                     Stopwatch readStopwatch = (Stopwatch) FileUtils.readFromFile(new File(PATH));
                     // Figure out the format by checking if the timer has been running for over an hour or not.
-                    String format;
-                    if (readStopwatch.getElapsedTime() >= 3600L) {
-                        format = "HH:mm:ss";
-                    } else {
-                        format = "mm:ss";
-                    }
-                    // Format the time with timezone UTC
-                    DateTimeFormatter formatter =
-                            DateTimeFormatter.ofPattern(format).withZone(ZoneId.of("UTC"));
 
-                    // Convert the time (long) to the format decided on above.
-                    Instant instant = Instant.ofEpochSecond(readStopwatch.getElapsedTime());
                     if (readStopwatch.isPaused()) {
                         // If the timer is paused, print the prefix and 'paused', if not print the prefix as well as the elapsed time.
-                        System.out.printf("%%{F%s}%s %s%%{F-}", PAUSED_COLOR, PREFIX, formatter.format(instant));
+                        System.out.printf("%%{F%s}%s %s%%{F-}", PAUSED_COLOR, prefix, readStopwatch.getFormattedElapsedTime());
                     } else {
-                        System.out.println(PREFIX + " " + formatter.format(instant));
+                        System.out.printf("%s %s", prefix, readStopwatch.getFormattedElapsedTime());
                     }
                 } else {
                     // If the timer hasn't been started print only the prefix icon.
-                    System.out.println(PREFIX);
+                    System.out.println(prefix);
                 }
             } else if (arguments.contains("stop") || arguments.contains("reset")) {
                 // If the argument is stop or reset, make sure the file is deleted on exit.
                 // This makes it so that if the program runs again it ends up printing only the prefix (see line 57)
                 File stopwatchFile = new File(PATH);
-                stopwatchFile.deleteOnExit();
+                if (stopwatchFile.exists()) {
+                    Stopwatch readStopwatch = (Stopwatch) FileUtils.readFromFile(stopwatchFile);
+                    // Check if the system the application is run on is running linux or not (requires dunst).
+                    if (OSUtils.isUnix()) {
+                        // Check if the --notif-icon argument is present.
+                        if (arguments.stream().anyMatch(arg -> arg.toLowerCase().contains("--notif-icon="))) {
+                            // Find the entry that specifies the notif icon path.
+                            arguments.forEach(arg -> {
+                                if (arg.toLowerCase().startsWith("--notif-icon=")) {
+                                    // Extract the path from the argument.
+                                    String iconPath = arg.substring(arg.indexOf("=") + 1);
+                                    try {
+                                        // Launch a dunst notification using dunstify, with urgency high and a timeout of 10 seconds.
+                                        Runtime.getRuntime().exec(new String[]{"dunstify", "Stopwatch stopped",
+                                                "Final elapsed time: " + readStopwatch.getFormattedElapsedTime(), "--timeout=10000",
+                                                "--raw_icon=" + iconPath});
+                                    } catch (IOException e) {
+                                        System.err.println("Unable to send notification.");
+                                        throw new RuntimeException(e);
+                                    }
+                                }
+                            });
+                        }
+
+                    }
+
+                    stopwatchFile.deleteOnExit();
+                }
             } else {
                 // Any other argument or an empty argument will end here. It will find the file and check if it exists.
                 File stopwatchFile = new File(PATH);
